Add first bit of code to calculate PCRs from tboot.
So far all I have is a class to parse an ACM per the documentation
in the "Intel® Trusted Execution Technology Measured Launched
Environment Developer’s Guide".  This is necessary to extract the
bits that are hashed as the initial parts of PCR 17.  The first hash
we need to compute is documented in section A.1.2 which requires
parsing the ACM as not every field is included in the hash.  See
http://www.mail-archive.com/tboot-devel@lists.sourceforge.net/msg00336.html
for details.