Don't add a newline when dumping the binary hash of a module.
Justification here is that we output binary when we want to feed the
hash into another program that would consider the newline in the hash.
An example would be piping the output of module-hash into the pcremu
program where it would simulate extending a PCR with the hash of
the module.